Move things out of MessageContext.cs
This file had no business containing all these random types.
